Moves logic away from the main app.py for readability
Fixes Santhoshkumard11/Flask-Calculator-Beginner-Docker#5
