+++ logic.py
@@ -1,0 +1,150 @@
+from flask_restful import Resource
+from flask import request, jsonify
+
+
+## Checks whether the posted data is valid
+def check_posted_data(postedData, functionName):
+    if (functionName == "add" or functionName == "subtract" or functionName == "multiply"):
+        if "x" not in postedData or "y" not in postedData:
+            return 301 #Missing parameter
+        else:
+            return 200
+    elif (functionName == "division"):
+        if "x" not in postedData or "y" not in postedData:
+            return 301
+        elif int(postedData["y"])==0:
+            return 302
+        else:
+            return 200
+
+
+# Math functions
+class Add(Resource):
+    def post(self):
+        #If I am here, then the resouce Add was requested using the method POST
+
+        #Step 1: Get posted data:
+        postedData = request.get_json()
+
+        print(postedData)
+
+        #Steb 1b: Verify validity of posted data
+        status_code = check_posted_data(postedData, "add")
+        if (status_code!=200):
+            retJson = {
+                "Message": "An error happened",
+                "Status Code":status_code
+            }
+            return jsonify(retJson)
+
+        #If i am here, then status_code == 200
+        x = postedData["x"]
+        y = postedData["y"]
+        x = int(x)
+        y = int(y)
+
+        #Step 2: Add the posted data
+        ret = x+y
+        retMap = {
+            'Message': ret,
+            'Status Code': 200
+        }
+        return jsonify(retMap)
+
+class Subtract(Resource):
+    def post(self):
+        #If I am here, then the resouce Subtract was requested using the method POST
+
+        #Step 1: Get posted data:
+        postedData = request.get_json()
+
+        #Steb 1b: Verify validity of posted data
+        status_code = check_posted_data(postedData, "subtract")
+
+
+        if (status_code!=200):
+            retJson = {
+                "Message": "An error happened",
+                "Status Code":status_code
+            }
+            return jsonify(retJson)
+
+        #If i am here, then status_code == 200
+        x = postedData["x"]
+        y = postedData["y"]
+        x = int(x)
+        y = int(y)
+
+        #Step 2: Subtract the posted data
+        ret = x-y
+        retMap = {
+            'Message': ret,
+            'Status Code': 200
+        }
+        return jsonify(retMap)
+
+
+class Multiply(Resource):
+    def post(self):
+        #If I am here, then the resouce Multiply was requested using the method POST
+
+        #Step 1: Get posted data:
+        postedData = request.get_json()
+
+        #Steb 1b: Verify validity of posted data
+        status_code = check_posted_data(postedData, "multiply")
+
+
+        if (status_code!=200):
+            retJson = {
+                "Message": "An error happened",
+                "Status Code":status_code
+            }
+            return jsonify(retJson)
+
+        #If i am here, then status_code == 200
+        x = postedData["x"]
+        y = postedData["y"]
+        x = int(x)
+        y = int(y)
+
+        #Step 2: Multiply the posted data
+        ret = x*y
+        retMap = {
+            'Message': ret,
+            'Status Code': 200
+        }
+        return jsonify(retMap)
+
+class Divide(Resource):
+    def post(self):
+        #If I am here, then the resouce Divide was requested using the method POST
+
+        #Step 1: Get posted data:
+        postedData = request.get_json()
+
+        #Steb 1b: Verify validity of posted data
+        status_code = check_posted_data(postedData, "division")
+
+
+        if (status_code!=200):
+            retJson = {
+                "Message": "An error happened",
+                "Status Code":status_code
+            }
+            return jsonify(retJson)
+
+        #If i am here, then status_code == 200
+        x = postedData["x"]
+        y = postedData["y"]
+        x = int(x)
+        y = int(y)
+
+        #Step 2: Multiply the posted data
+        ret = (x*1.0)/y
+        retMap = {
+            'Message': ret,
+            'Status Code': 200
+        }
+        return jsonify(retMap)
+
